Setup.

+++ Assets/Scripts/Interfaces/IMove.cs
@@ -12,7 +12,7 @@ namespace HMF.Thesis.Interfaces
         int BaseSpeed {get; set;}
 
         /// The current speed of the Character;
-        int Speed {get;set;}
+        float Speed {get;set;}
 
         /// The speed of the Jump.
         int JumpSpeed {get; set;}
